refactoring to separate doc based and aem based code
