+++ blocks/form/components/file/file.js
@@ -1,4 +1,4 @@
-import { updateOrCreateInvalidMsg, stripTags } from '../../util.js';
+import { updateOrCreateInvalidMsg, stripTags } from '../../lib/util.js';
 import { fileAttachmentText, dragDropText, defaultErrorMessages } from '../../constant.js';
 
 const fileSizeRegex = /^(\d*\.?\d+)(\\?(?=[KMGT])([KMGT])(?:i?B)?|B?)$/i;
