+++ blocks/form/components/wizard/wizard.js
@@ -1,4 +1,4 @@
-import { createButton } from '../../util.js';
+import { createButton } from '../../lib/util.js';
 
 export class WizardLayout {
   inputFields = 'input,textarea,select';
