+++ blocks/form/constant.js
@@ -18,7 +18,6 @@ export const defaultErrorMessages = {
 
 // eslint-disable-next-line no-useless-escape
 export const emailPattern = '([A-Za-z0-9][._]?)+[A-Za-z0-9]@[A-Za-z0-9]+(\.?[A-Za-z0-9]){2}\.([A-Za-z0-9]{2,4})?';
-
 let submitBaseUrl = '';
 
 export function setSubmitBaseUrl(url) {
