+++ blocks/form/lib/transform.js
@@ -191,7 +191,7 @@ export default class DocBasedFormToAF {
      *
      * @return {{formDef: any, excelData: any}} response
      */
-  transform(exData, { name } = { name: 'Form' }) {
+  transform(exData, path, { name } = { name: 'Form' }) {
     this.errors = [];
     // if its adaptive form json just return it.
     if (exData?.adaptiveform) {
@@ -237,6 +237,7 @@ export default class DocBasedFormToAF {
       }
     });
     formDef.properties.rules = { fieldIdMap, rules };
+    formDef.action = path?.split('.json')[0];
     return formDef;
   }
