+++ blocks/form/lib/util.js
@@ -1,5 +1,5 @@
 // create a string containing head tags from h1 to h5
-import { defaultErrorMessages } from './constant.js';
+import { defaultErrorMessages } from '../constant.js';
 
 const headings = Array.from({ length: 5 }, (_, i) => `<h${i + 1}>`).join('');
 const allowedTags = `${headings}<a><b><p><i><em><strong><ul><li><ol>`;
