+++ blocks/form/submit.js
@@ -1,21 +1,22 @@
 import { DEFAULT_THANK_YOU_MESSAGE } from './constant.js';
 
+// eslint-disable-next-line no-unused-vars
 export function submitSuccess(e, form) {
   const { payload } = e;
-  const redirectUrl = form.dataset.redirectUrl || payload?.body?.redirectUrl;
-  const thankYouMsg = form.dataset.thankYouMsg || payload?.body?.thankYouMessage;
+  const redirectUrl = payload?.body?.redirectUrl;
+  const thankYouMessage = payload?.body?.thankYouMessage;
   if (redirectUrl) {
     window.location.assign(encodeURI(redirectUrl));
   } else {
-    let thankYouMessage = form.parentNode.querySelector('.form-message.success-message');
-    if (!thankYouMessage) {
-      thankYouMessage = document.createElement('div');
-      thankYouMessage.className = 'form-message success-message';
+    let thankYouMsgEl = form.parentNode.querySelector('.form-message.success-message');
+    if (!thankYouMsgEl) {
+      thankYouMsgEl = document.createElement('div');
+      thankYouMsgEl.className = 'form-message success-message';
     }
-    thankYouMessage.innerHTML = thankYouMsg || DEFAULT_THANK_YOU_MESSAGE;
-    form.parentNode.insertBefore(thankYouMessage, form);
-    if (thankYouMessage.scrollIntoView) {
-      thankYouMessage.scrollIntoView({ behavior: 'smooth' });
+    thankYouMsgEl.innerHTML = thankYouMessage || DEFAULT_THANK_YOU_MESSAGE;
+    form.parentNode.insertBefore(thankYouMsgEl, form);
+    if (thankYouMsgEl.scrollIntoView) {
+      thankYouMsgEl.scrollIntoView({ behavior: 'smooth' });
     }
     form.reset();
   }
@@ -23,6 +24,7 @@ export function submitSuccess(e, form) {
   form.querySelector('button[type="submit"]').disabled = false;
 }
 
+// eslint-disable-next-line no-unused-vars
 export function submitFailure(e, form) {
   let errorMessage = form.querySelector('.form-message.error-message');
   if (!errorMessage) {
@@ -34,97 +36,4 @@ export function submitFailure(e, form) {
   errorMessage.scrollIntoView({ behavior: 'smooth' });
   form.setAttribute('data-submitting', 'false');
   form.querySelector('button[type="submit"]').disabled = false;
-}
-
-function generateUnique() {
-  return new Date().valueOf() + Math.random();
-}
-
-function getFieldValue(fe, payload) {
-  if (fe.type === 'radio') {
-    return fe.form.elements[fe.name].value;
-  } if (fe.type === 'checkbox') {
-    if (fe.checked) {
-      if (payload[fe.name]) {
-        return `${payload[fe.name]},${fe.value}`;
-      }
-      return fe.value;
-    }
-  } else if (fe.type !== 'file') {
-    return fe.value;
-  }
-  return null;
-}
-
-function constructPayload(form) {
-  const payload = { __id__: generateUnique() };
-  [...form.elements].forEach((fe) => {
-    if (fe.name && !fe.matches('button') && !fe.disabled && fe.tagName !== 'FIELDSET') {
-      const value = getFieldValue(fe, payload);
-      if (fe.closest('.repeat-wrapper')) {
-        payload[fe.name] = payload[fe.name] ? `${payload[fe.name]},${fe.value}` : value;
-      } else {
-        payload[fe.name] = value;
-      }
-    }
-  });
-  return { payload };
-}
-
-async function prepareRequest(form) {
-  const { payload } = constructPayload(form);
-  const headers = {
-    'Content-Type': 'application/json',
-  };
-  const body = { data: payload };
-  const url = form.dataset.submit || form.dataset.action;
-  return { headers, body, url };
-}
-
-async function submitDocBasedForm(form, captcha) {
-  try {
-    const { headers, body, url } = await prepareRequest(form, captcha);
-    let token = null;
-    if (captcha) {
-      token = await captcha.getToken();
-      body.data['g-recaptcha-response'] = token;
-    }
-    const response = await fetch(url, {
-      method: 'POST',
-      headers,
-      body: JSON.stringify(body),
-    });
-    if (response.ok) {
-      submitSuccess(response, form);
-    } else {
-      const error = await response.text();
-      throw new Error(error);
-    }
-  } catch (error) {
-    submitFailure(error, form);
-  }
-}
-
-export async function handleSubmit(e, form, captcha) {
-  e.preventDefault();
-  const valid = form.checkValidity();
-  if (valid) {
-    e.submitter?.setAttribute('disabled', '');
-    if (form.getAttribute('data-submitting') !== 'true') {
-      form.setAttribute('data-submitting', 'true');
-
-      // hide error message in case it was shown before
-      form.querySelectorAll('.form-message.show').forEach((el) => el.classList.remove('show'));
-
-      if (form.dataset.source === 'sheet') {
-        await submitDocBasedForm(form, captcha);
-      }
-    }
-  } else {
-    const firstInvalidEl = form.querySelector(':invalid:not(fieldset)');
-    if (firstInvalidEl) {
-      firstInvalidEl.focus();
-      firstInvalidEl.scrollIntoView({ behavior: 'smooth' });
-    }
-  }
 }
