+++ package.json
@@ -5,10 +5,10 @@
   "type": "module",
   "c8": {
     "exclude": [
-      "blocks/form/rules/model/*.js",
-      "blocks/form/rules/formula/*.js",
-      "blocks/form/rules/worker.js",
-      "blocks/form/rules-doc/parser/*.js"
+      "blocks/form/lib/rules/model/*.js",
+      "blocks/form/lib/rules/formula/*.js",
+      "blocks/form/lib/rules/worker.js",
+      "blocks/form/lib/rules-doc/parser/*.js"
     ],
     "include": [
       "blocks/form/**/*.js",
@@ -58,7 +58,7 @@
     "@aemforms/af-formatters": "0.22.106",
     "@babel/core": "7.24.0",
     "@babel/eslint-parser": "7.23.10",
-    "@playwright/test": "^1.43.1",
+    "@playwright/test": "^1.47.2",
     "@testing-library/dom": "^9.3.4",
     "@testing-library/user-event": "14.3.0",
     "@types/node": "^20.12.7",
@@ -68,13 +68,17 @@
     "eslint": "8.57.0",
     "eslint-config-airbnb-base": "15.0.0",
     "eslint-plugin-import": "2.29.1",
+    "eslint-plugin-json": "3.1.0",
     "fs-extra": "^11.2.0",
+    "husky": "9.1.1",
     "jsdom": "^24.0.0",
     "jsdom-global": "^3.0.2",
+    "merge-json-cli": "1.0.3",
     "mocha": "^10.2.0",
     "mutationobserver-shim": "^0.3.7",
     "nock": "^13.4.0",
     "node-fetch": "^3.3.2",
+    "npm-run-all": "4.1.5",
     "nyc": "^15.1.0",
     "parse-multipart-data": "^1.5.0",
     "rollup": "^2.79.1",
@@ -82,10 +86,6 @@
     "rollup-plugin-license": "^3.0.1",
     "rollup-plugin-terser": "^7.0.2",
     "sinon": "^17.0.1",
-    "eslint-plugin-json": "3.1.0",
-    "husky": "9.1.1",
-    "merge-json-cli": "1.0.3",
-    "npm-run-all": "4.1.5",
     "stylelint": "16.6.1",
     "stylelint-config-standard": "36.0.0"
   },
