+++ test/unit/fixtures/form/tripPlanner.js
@@ -32,4 +32,10 @@ export const fieldDef = {
     Name: 'attach', Type: 'file', Placeholder: '', Label: 'Attach', Mandatory: '', Value: '', Visible: '', Min: '', Max: '1MB', Fieldset: '', Repeatable: '', Options: '', OptionNames: '', 'Value Expression': '', 'Visible Expression': '', 'Column Span': '', Checked: '', Step: '', ReadOnly: '', Description: '', 'Required Error Message': '', Accept: 'application/pdf',
   }],
   ':type': 'sheet',
+  action: '/adobe/forms/af/form1',
+  id: 'form',
 };
+
+export const bUrlMode = true;
+
+export const formPath = 'http://localhost:3000/adobe/forms/af/form1.json';
